refactor(adapters,mutator): extract columns, action ordering, action creation
Extract focused modules from the two largest remaining files:

adapters/files.lua (717 -> 467 lines, -35%):
- files/columns.lua (267 lines): column definitions (size, owner, group,
  permissions, time), columns_require_stat, get_column. Includes all
  platform-conditional logic (ids, permissions).

mutator/init.lua (706 -> 273 lines, -61%):
- mutator/action_ordering.lua (191 lines): topological sort with cycle
  detection, move loop splitting, dependency graph computation via Trie.
- mutator/action_creation.lua (265 lines): diff-to-action conversion,
  path expansion (extglob), create deduplication, type annotations.

Thin re-export wrappers on parent modules preserve backwards compatibility.
All type annotations (canola.Action et al.) moved to action_creation.lua.
