+++ lua/canola/mutator/init.lua
@@ -1,11 +1,11 @@
 local Progress = require('canola.mutator.progress')
-local Trie = require('canola.mutator.trie')
+local action_creation = require('canola.mutator.action_creation')
+local action_ordering = require('canola.mutator.action_ordering')
 local cache = require('canola.cache')
 local canola = require('canola')
 local columns = require('canola.columns')
 local config = require('canola.config')
 local confirmation = require('canola.mutator.confirmation')
-local constants = require('canola.constants')
 local fs = require('canola.fs')
 local lsp_helpers = require('canola.lsp.helpers')
 local parser = require('canola.mutator.parser')
@@ -13,445 +13,12 @@ local util = require('canola.util')
 local view = require('canola.view')
 local M = {}
 
-local FIELD_NAME = constants.FIELD_NAME
-local FIELD_TYPE = constants.FIELD_TYPE
-
----@alias canola.Action canola.CreateAction|canola.DeleteAction|canola.MoveAction|canola.CopyAction|canola.ChangeAction
-
----@class (exact) canola.CreateAction
----@field type "create"
----@field url string
----@field entry_type canola.EntryType
----@field link nil|string
-
----@class (exact) canola.DeleteAction
----@field type "delete"
----@field url string
----@field entry_type canola.EntryType
-
----@class (exact) canola.MoveAction
----@field type "move"
----@field entry_type canola.EntryType
----@field src_url string
----@field dest_url string
-
----@class (exact) canola.CopyAction
----@field type "copy"
----@field entry_type canola.EntryType
----@field src_url string
----@field dest_url string
-
----@class (exact) canola.ChangeAction
----@field type "change"
----@field entry_type canola.EntryType
----@field url string
----@field column string
----@field value any
-
-local EXTGLOB_HARD_CAP = 10000
-
----@param name string
----@return string[][]
-local function expand_path_segments(name)
-  local extglob_setting = config.extglob
-  local path_sep = fs.is_windows and '[/\\]' or '/'
-  local segments = vim.split(name, path_sep)
-
-  if extglob_setting == false then
-    return { segments }
-  end
-
-  local extglob = require('canola.extglob')
-  local expanded_segments = {}
-  for _, seg in ipairs(segments) do
-    table.insert(expanded_segments, extglob.expand(seg))
-  end
-
-  local results = { {} }
-  for _, alternatives in ipairs(expanded_segments) do
-    local new_results = {}
-    for _, partial in ipairs(results) do
-      for _, alt in ipairs(alternatives) do
-        local new = vim.list_extend({}, partial)
-        table.insert(new, alt)
-        table.insert(new_results, new)
-      end
-    end
-    results = new_results
-    if #results > EXTGLOB_HARD_CAP then
-      vim.notify(
-        string.format(
-          '[canola] Brace expansion exceeds hard cap of %d entries, aborting',
-          EXTGLOB_HARD_CAP
-        ),
-        vim.log.levels.ERROR
-      )
-      return { segments }
-    end
-  end
-
-  if type(extglob_setting) == 'number' and #results > extglob_setting then
-    local choice = vim.fn.confirm(
-      string.format('Brace expansion will create %d entries. Continue?', #results),
-      '&Yes\n&No',
-      2
-    )
-    if choice ~= 1 then
-      return { segments }
-    end
-  end
-
-  return results
-end
-
----@param all_diffs table<integer, canola.Diff[]>
----@return canola.Action[]
 M.create_actions_from_diffs = function(all_diffs)
-  ---@type canola.Action[]
-  local actions = {}
-
-  ---@type table<integer, canola.Diff[]>
-  local diff_by_id = setmetatable({}, {
-    __index = function(t, key)
-      local list = {}
-      rawset(t, key, list)
-      return list
-    end,
-  })
-
-  -- To deduplicate create actions
-  -- This can happen when creating deep nested files e.g.
-  -- > foo/bar/a.txt
-  -- > foo/bar/b.txt
-  local seen_creates = {}
-
-  ---@param action canola.Action
-  local function add_action(action)
-    local adapter = assert(config.get_adapter_by_scheme(action.dest_url or action.url))
-    if not adapter.filter_action or adapter.filter_action(action) then
-      if action.type == 'create' then
-        if seen_creates[action.url] then
-          return
-        else
-          seen_creates[action.url] = true
-        end
-      end
-
-      table.insert(actions, action)
-    end
-  end
-  for bufnr, diffs in pairs(all_diffs) do
-    local adapter = util.get_adapter(bufnr, true)
-    if not adapter then
-      error('Missing adapter')
-    end
-    local parent_url = vim.api.nvim_buf_get_name(bufnr)
-    for _, diff in ipairs(diffs) do
-      if diff.type == 'new' then
-        if diff.id then
-          local expanded_names = expand_path_segments(diff.name)
-          for _, segments in ipairs(expanded_names) do
-            local expanded_name = table.concat(segments, '/')
-            local cloned = vim.deepcopy(diff)
-            local by_id = diff_by_id[diff.id]
-            ---HACK: set the destination on this diff for use later
-            ---@diagnostic disable-next-line: inject-field
-            cloned.dest = parent_url .. expanded_name
-            ---@diagnostic disable-next-line: inject-field
-            cloned._segments = segments
-            ---@diagnostic disable-next-line: inject-field
-            cloned._parent_url = parent_url
-            table.insert(by_id, cloned)
-          end
-        else
-          local expanded_paths = expand_path_segments(diff.name)
-          for _, segments in ipairs(expanded_paths) do
-            local url = parent_url:gsub('/$', '')
-            for i, seg in ipairs(segments) do
-              local is_last = i == #segments
-              local entry_type = is_last and diff.entry_type or 'directory'
-              url = url .. '/' .. seg
-              add_action({
-                type = 'create',
-                url = url,
-                entry_type = entry_type,
-                link = is_last and diff.link or nil,
-              })
-            end
-          end
-        end
-      elseif diff.type == 'change' then
-        add_action({
-          type = 'change',
-          url = parent_url .. diff.name,
-          entry_type = diff.entry_type,
-          column = diff.column,
-          value = diff.value,
-        })
-      else
-        local by_id = diff_by_id[diff.id]
-        -- HACK: set has_delete field on a list-like table of diffs
-        ---@diagnostic disable-next-line: inject-field
-        by_id.has_delete = true
-        -- Don't insert the delete. We already know that there is a delete because of the presence
-        -- in the diff_by_id map. The list will only include the 'new' diffs.
-      end
-    end
-  end
-
-  for id, diffs in pairs(diff_by_id) do
-    local entry = cache.get_entry_by_id(id)
-    if not entry then
-      error(string.format('Could not find entry %d', id))
-    end
-    ---HACK: access the has_delete field on the list-like table of diffs
-    ---@diagnostic disable-next-line: undefined-field
-    if diffs.has_delete then
-      local has_create = #diffs > 0
-      if has_create then
-        -- MOVE (+ optional copies) when has both creates and delete
-        for i, diff in ipairs(diffs) do
-          ---@diagnostic disable-next-line: undefined-field
-          if diff._segments and #diff._segments > 1 then
-            ---@diagnostic disable-next-line: undefined-field
-            local url = diff._parent_url:gsub('/$', '')
-            ---@diagnostic disable-next-line: undefined-field
-            for j = 1, #diff._segments - 1 do
-              ---@diagnostic disable-next-line: undefined-field
-              url = url .. '/' .. diff._segments[j]
-              add_action({ type = 'create', url = url, entry_type = 'directory' })
-            end
-          end
-          add_action({
-            type = i == #diffs and 'move' or 'copy',
-            entry_type = entry[FIELD_TYPE],
-            ---HACK: access the dest field we set above
-            ---@diagnostic disable-next-line: undefined-field
-            dest_url = diff.dest,
-            src_url = cache.get_parent_url(id) .. entry[FIELD_NAME],
-          })
-        end
-      else
-        -- DELETE when no create
-        add_action({
-          type = 'delete',
-          entry_type = entry[FIELD_TYPE],
-          url = cache.get_parent_url(id) .. entry[FIELD_NAME],
-        })
-      end
-    else
-      -- COPY when create but no delete
-      for _, diff in ipairs(diffs) do
-        ---@diagnostic disable-next-line: undefined-field
-        if diff._segments and #diff._segments > 1 then
-          ---@diagnostic disable-next-line: undefined-field
-          local url = diff._parent_url:gsub('/$', '')
-          ---@diagnostic disable-next-line: undefined-field
-          for j = 1, #diff._segments - 1 do
-            ---@diagnostic disable-next-line: undefined-field
-            url = url .. '/' .. diff._segments[j]
-            add_action({ type = 'create', url = url, entry_type = 'directory' })
-          end
-        end
-        add_action({
-          type = 'copy',
-          entry_type = entry[FIELD_TYPE],
-          src_url = cache.get_parent_url(id) .. entry[FIELD_NAME],
-          ---HACK: access the dest field we set above
-          ---@diagnostic disable-next-line: undefined-field
-          dest_url = diff.dest,
-        })
-      end
-    end
-  end
-
-  return M.enforce_action_order(actions)
+  return action_creation.create_actions_from_diffs(all_diffs)
 end
 
----@param actions canola.Action[]
----@return canola.Action[]
 M.enforce_action_order = function(actions)
-  local src_trie = Trie.new()
-  local dest_trie = Trie.new()
-  for _, action in ipairs(actions) do
-    if action.type == 'delete' or action.type == 'change' then
-      src_trie:insert_action(action.url, action)
-    elseif action.type == 'create' then
-      dest_trie:insert_action(action.url, action)
-    else
-      dest_trie:insert_action(action.dest_url, action)
-      src_trie:insert_action(action.src_url, action)
-    end
-  end
-
-  -- 1. create a graph, each node points to all of its dependencies
-  -- 2. for each action, if not added, find it in the graph
-  -- 3. traverse through the graph until you reach a node that has no dependencies (leaf)
-  -- 4. append that action to the return value, and remove it from the graph
-  --   a. TODO optimization: check immediate parents to see if they have no dependencies now
-  -- 5. repeat
-
-  ---Gets the dependencies of a particular action. Effectively dynamically calculates the dependency
-  ---"edges" of the graph.
-  ---@param action canola.Action
-  local function get_deps(action)
-    local ret = {}
-    if action.type == 'delete' then
-      src_trie:accum_children_of(action.url, ret)
-    elseif action.type == 'create' then
-      -- Finish operating on parents first
-      -- e.g. NEW /a BEFORE NEW /a/b
-      dest_trie:accum_first_parents_of(action.url, ret)
-      -- Process remove path before creating new path
-      -- e.g. DELETE /a BEFORE NEW /a
-      src_trie:accum_actions_at(action.url, ret, function(a)
-        return a.type == 'move' or a.type == 'delete'
-      end)
-    elseif action.type == 'change' then
-      -- Finish operating on parents first
-      -- e.g. NEW /a BEFORE CHANGE /a/b
-      dest_trie:accum_first_parents_of(action.url, ret)
-      -- Finish operations on this path first
-      -- e.g. NEW /a BEFORE CHANGE /a
-      dest_trie:accum_actions_at(action.url, ret)
-      -- Finish copy from operations first
-      -- e.g. COPY /a -> /b BEFORE CHANGE /a
-      src_trie:accum_actions_at(action.url, ret, function(entry)
-        return entry.type == 'copy'
-      end)
-    elseif action.type == 'move' then
-      -- Finish operating on parents first
-      -- e.g. NEW /a BEFORE MOVE /z -> /a/b
-      dest_trie:accum_first_parents_of(action.dest_url, ret)
-      -- Process children before moving
-      -- e.g. NEW /a/b BEFORE MOVE /a -> /b
-      dest_trie:accum_children_of(action.src_url, ret)
-      -- Process children before moving parent dir
-      -- e.g. COPY /a/b -> /b BEFORE MOVE /a -> /d
-      -- e.g. CHANGE /a/b BEFORE MOVE /a -> /d
-      src_trie:accum_children_of(action.src_url, ret)
-      -- Process remove path before moving to new path
-      -- e.g. MOVE /a -> /b BEFORE MOVE /c -> /a
-      src_trie:accum_actions_at(action.dest_url, ret, function(a)
-        return a.type == 'move' or a.type == 'delete'
-      end)
-    elseif action.type == 'copy' then
-      -- Finish operating on parents first
-      -- e.g. NEW /a BEFORE COPY /z -> /a/b
-      dest_trie:accum_first_parents_of(action.dest_url, ret)
-      -- Process children before copying
-      -- e.g. NEW /a/b BEFORE COPY /a -> /b
-      dest_trie:accum_children_of(action.src_url, ret)
-      -- Process remove path before copying to new path
-      -- e.g. MOVE /a -> /b BEFORE COPY /c -> /a
-      src_trie:accum_actions_at(action.dest_url, ret, function(a)
-        return a.type == 'move' or a.type == 'delete'
-      end)
-    end
-    return ret
-  end
-
-  ---@return nil|canola.Action The leaf action
-  ---@return nil|canola.Action When no leaves found, this is the last action in the loop
-  local function find_leaf(action, seen)
-    if not seen then
-      seen = {}
-    elseif seen[action] then
-      return nil, action
-    end
-    seen[action] = true
-    local deps = get_deps(action)
-    if next(deps) == nil then
-      return action
-    end
-    local action_in_loop
-    for _, dep in ipairs(deps) do
-      local leaf, loop_action = find_leaf(dep, seen)
-      if leaf then
-        return leaf
-      elseif not action_in_loop and loop_action then
-        action_in_loop = loop_action
-      end
-    end
-    return nil, action_in_loop
-  end
-
-  local ret = {}
-  local after = {}
-  while next(actions) ~= nil do
-    local action = actions[1]
-    local selected, loop_action = find_leaf(action)
-    local to_remove
-    if selected then
-      to_remove = selected
-    else
-      if loop_action and loop_action.type == 'move' then
-        -- If this is moving a parent into itself, that's an error
-        if vim.startswith(loop_action.dest_url, loop_action.src_url) then
-          error('Detected cycle in desired paths')
-        end
-
-        -- We've detected a move cycle (e.g. MOVE /a -> /b + MOVE /b -> /a)
-        -- Split one of the moves and retry
-        local intermediate_url =
-          string.format('%s__oil_tmp_%05d', loop_action.src_url, math.random(999999))
-        local move_1 = {
-          type = 'move',
-          entry_type = loop_action.entry_type,
-          src_url = loop_action.src_url,
-          dest_url = intermediate_url,
-        }
-        local move_2 = {
-          type = 'move',
-          entry_type = loop_action.entry_type,
-          src_url = intermediate_url,
-          dest_url = loop_action.dest_url,
-        }
-        to_remove = loop_action
-        table.insert(actions, move_1)
-        table.insert(after, move_2)
-        dest_trie:insert_action(move_1.dest_url, move_1)
-        src_trie:insert_action(move_1.src_url, move_1)
-      else
-        error('Detected cycle in desired paths')
-      end
-    end
-
-    if selected then
-      if selected.type == 'move' or selected.type == 'copy' then
-        if vim.startswith(selected.dest_url, selected.src_url .. '/') then
-          error(
-            string.format(
-              'Cannot move or copy parent into itself: %s -> %s',
-              selected.src_url,
-              selected.dest_url
-            )
-          )
-        end
-      end
-      table.insert(ret, selected)
-    end
-
-    if to_remove then
-      if to_remove.type == 'delete' or to_remove.type == 'change' then
-        src_trie:remove_action(to_remove.url, to_remove)
-      elseif to_remove.type == 'create' then
-        dest_trie:remove_action(to_remove.url, to_remove)
-      else
-        dest_trie:remove_action(to_remove.dest_url, to_remove)
-        src_trie:remove_action(to_remove.src_url, to_remove)
-      end
-      for i, a in ipairs(actions) do
-        if a == to_remove then
-          table.remove(actions, i)
-          break
-        end
-      end
-    end
-  end
-
-  vim.list_extend(ret, after)
-  return ret
+  return action_ordering.enforce_action_order(actions)
 end
 
 local progress
